refactor(providers): extract actions modules for hackernews, a_leads, abstract
These three providers inlined their action definitions in definition.ts
instead of the standard definition.ts / actions.ts / executors.ts layout
used by the rest of the catalog. Move the action definitions and their
supporting schemas into provider actions.ts modules (hackernewsActions,
aLeadsActions, abstractActions) and keep definition.ts as catalog
metadata only.

Pure move: generated catalog JSON for all three providers is
byte-identical before and after.
